Fix a build warning, patch by nobled!
git-svn-id: https://llvm.org/svn/llvm-project/klee/trunk@115315 91177308-0d34-0410-b5e6-96231b3b80d8

+++ tools/klee-replay/klee-replay.c
@@ -362,7 +362,7 @@ void klee_make_symbolic(void *addr, size_t nbytes, const char *name) {
     } else {
       if (boo->numBytes != nbytes) {
 	fprintf(stderr, "make_symbolic mismatch, different sizes: "
-		"%d in input file, %ld in code\n", boo->numBytes, nbytes);
+		"%d in input file, %lu in code\n", boo->numBytes, (unsigned long)nbytes);
 	exit(1);
       } else {
         memcpy(addr, boo->bytes, nbytes);
